Fix two bugs in AtTimer
Bug #1: if hour/min/sec sequences weren't specified in order, some times
would be missed.

Bug #2: the timer could improperly fire at incorrect times; namely if a minute
later than the current minute was specified, the timer would fire on the
current hour at that minute even if the current hour was not specified in the
hours list.  Ditto for seconds and the current hour/minute.


git-svn-id: svn://svn.freevo.org/kaa/trunk/base@4641 a8f5125c-1e01-0410-8897-facf34644b8e