Python 3: new strutils fsname() and bl() functions
fsname() returns an object appropriate to represent a filename for the current
Python version.  On Python 2 a non-unicode string, and on Python 3, a unicode
string using the surrogateescape handler.  See PEP 383.

bl() is a transition function while we must still support Python 2.5, and returns
a bytes literal.  bl('foo') is equivalent to b'foo', albeit slower due to a
function call.



git-svn-id: svn://svn.freevo.org/kaa/trunk/base@4674 a8f5125c-1e01-0410-8897-facf34644b8e