Python 3: fix kaa.db to work between Python 2 and 3
This is complicated by the fact that pickles between Python 2 and Python 3
aren't _really_ compatible.  There are differences between the way string
types are handled.  We use a custom protocol 2 unpickler for python 3
that loads pickled strings as bytes unless they're used as dictionary keys.

Currently databases created by Python 2 and read in Python 3 are readonly
until the to-be-implemented upgrade_to_py3() method is called, which will
convert all proto2 pickles to proto3, which understands the bytes type.



git-svn-id: svn://svn.freevo.org/kaa/trunk/base@4689 a8f5125c-1e01-0410-8897-facf34644b8e