Resume coroutine immediately when a dependent InProgress finishes
Previously, the coroutine timer was restarted when an InProgress dependency
was finished.  This involved stepping back into the main loop, and doing a
select() with the coroutine's interval as the select() timeout.

With this change, coroutines with 0 intervals will resume immediately when
a dependent InProgress finishes.  This halves the number of select syscalls
when coroutines yield Socket.read() or Socket.write()