Put stop() in a OneShotTimer to mitigate possible deadlocks
Originally, stop() was decorated with @threaded(MAINTHREAD).  We take this one
step further by decorating it with @timed(policy=POLICY_ONCE) so that even if
stop() is invoked from the main thread, we wait until the current iteration is
finished before calling all the shutdown handlers.

This fixes at least one corner case deadlock, for example if stop() is called
from an asynchronous function like a signal handler.