Improve SIGCHLD handling
Remove the poorly thought-out 'unix-signal' kaa signal and replace it
with SIGCHLD.  A dispatcher for generic POSIX signal handling was adding
for CoreThreading and it can later be used for other signals.

Perhaps most importantly, this fixes a regression where children spawned
by Process() objects were not getting reaped, because of the use of a
compound expression in a lambda:

   notifier.socket_add(pipe[0], lambda fd: CoreThreading._purge_pipe(fd) and signals['unix-signal'].emit())

If _purge_pipe() returns a false value (which it did, None) then
'unix-signal' never gets emitted.  This prevented children from getting
reaped and Process objects to live forever.