Do not consider channels in the process of closing as readable
Fixes a problem where this loop inside a coroutine may never terminate:

   while channel.readable:
       yield channel.read()

Here, if IOChannel internally reads the empty string from the channel,
it is considered EOF and closed with close(immediate=True).  close() will
finish any pending read(), which resumes the coroutine running the above loop
and causes channel.readable to be evaluated again.

Because this evaluation happens in the middle of close() before channel is set
to None, and the readable property did not test _closing, channel.readable was
True, when it should have been False to terminate the loop.