Improve InProgress abortability
Allow abortable property to be explicitly set False, which will prevent
abort() even if there are callbacks connected to the 'abort' signal.

If an InProgressAborted is thrown to the InProgress, don't raise the
'abort' signal unless the InProgress is abortable.

Allow InProgress returned by timeout() to be abort()ed, and doing so will
abort the original InProgress.

If timeout(abort=True) is called, eat any InProgressAborted exception raised
by abort(), because the exception is superfluous.

Introduce new InProgress.noabort() which creates a new InProgress linked with
the original that can't be aborted.