Improve abortability of coroutines
Raise InProgressAborted inside the generator instead of the ambiguous
GeneratorExit, which allows coroutines to handle the case when an
InProgress the coroutine has yielded is aborted.

Also, if a coroutine is aborted, do not attempt to abort a yielded InProgress
which is not abortable, or if something else is waiting for the yielded
InProgress's value.