Add option for close behaviour and lock read queue
By default, IOChannel will close when it encounters EOF, which is sensible for
streams, but not for files.  New close_on_eof property controls this behaviour.

Also the read queue is now managed by an RLock, which fixes some raciness with
concurrent access between threads.

Finally, allow close() to be invoked from a thread, and return an InProgress
which finishes once the channel is closed.